// src/server/api/routers/smartsheet/smartsheetTypes.ts
export enum AccessLevel {
  ADMIN = "ADMIN",
  COMMENTER = "COMMENTER",
  EDITOR = "EDITOR",
  EDITOR_SHARE = "EDITOR_SHARE",
  OWNER = "OWNER",
  VIEWER = "VIEWER",
}

export enum AttachmentType {
  BOX_COM = "BOX_COM",
  DROPBOX = "DROPBOX",
  EGNYTE = "EGNYTE",
  FILE = "FILE",
  GOOGLE_DRIVE = "GOOGLE_DRIVE",
  LINK = "LINK",
  ONEDRIVE = "ONEDRIVE",
  SMARTSHEET = "SMARTSHEET",
}

export enum AttachmentSubType {
  DOCUMENT = "DOCUMENT",
  DRAWING = "DRAWING",
  FOLDER = "FOLDER",
  PDF = "PDF",
  PRESENTATION = "PRESENTATION",
  SPREADSHEET = "SPREADSHEET",
}

export enum ColumnTypes {
  ABSTRACT_DATETIME = "ABSTRACT_DATETIME",
  CHECKBOX = "CHECKBOX",
  CONTACT_LIST = "CONTACT_LIST",
  DATE = "DATE",
  DATETIME = "DATETIME",
  DURATION = "DURATION",
  MULTI_CONTACT_LIST = "MULTI_CONTACT_LIST",
  MULTI_PICKLIST = "MULTI_PICKLIST",
  PICKLIST = "PICKLIST",
  PREDECESSOR = "PREDECESSOR",
  TEXT_NUMBER = "TEXT_NUMBER",
}
// ...
export enum Status {
  BLOCKED = "BLOCKED",
  BROKEN = "BROKEN",
  CIRCULAR = "CIRCULAR",
  DISABLED = "DISABLED",
  INACCESSIBLE = "INACCESSIBLE",
  INVALID = "INVALID",
  NOT_SHARED = "NOT_SHARED",
  OK = "OK",
}

export enum SummaryPermissions {
  ADMIN = "ADMIN",
  READ_DELETE = "READ_DELETE",
  READ_ONLY = "READ_ONLY",
  READ_WRITE = "READ_WRITE",
}

export enum SystemColumns {
  AUTO_NUMBER = "AUTO_NUMBER",
  CREATED_BY = "CREATED_BY",
  CREATED_DATE = "CREATED_DATE",
  MODIFIED_BY = "MODIFIED_BY",
  MODIFIED_DATE = "MODIFIED_DATE",
}

export enum Tags {
  CALENDAR_END_DATE = "CALENDAR_END_DATE",
  CALENDAR_START_DATE = "CALENDAR_START_DATE",
  CARD_DONE = "CARD_DONE",
  GANTT_ALLOCATION = "GANNTT_ALLOCATION",
  GANTT_ASSIGNED_RESOURCE = "GANTT_ASSIGNED_RESOURCE",
  GANTT_DISPLAY_LABEL = "GANTT_DISPLAY_LABEL",
  GANTT_DURATION = "GANTT_DURATION",
  GANTT_END_DATE = "GANTT_END_DATE",
  GANTT_PERCENTAGE_COMPLETE = "GANTT_PERCENTAGE_COMPLETE",
  GANTT_PREDECESSOR = "GANTT_PREDECESSOR",
  GANTT_START_DATE = "GANTT_START_DATE",
  BASELINE_START_DATE = "BASELINE_START_DATE",
  BASELINE_END_DATE = "BASELINE_END_DATE",
  BASELINE_VARIANCE = "BASELINE_VARIANCE",
}
// ...
export type Attachment = {
  id: number;
  parentId: number;
  attachmentType: AttachmentType;
  attachmetSubType: AttachmentSubType;
  mimeType: string;
  parentType: string;
  createdAt: string;
  createdBy: User;
  name: string;
  sizeInKb: number;
  url: string;
  urlExpiresInMillis: number;
};

export type Axis = {
  location: Locations;
  title: string;
  titleInfo: ColumnTypes;
  axisLabelFormat: string;
  includeZero: boolean;
};
export type AutoNumberFormat = {
  fill: string;
  prefix: string;
  startingNumber: number;
  suffix: string;
};
// ...
export type User = {
  id: string;
  email: string | null;
  name: string | null;
};

export type UserPermissions = {
  summaryPermissions: SummaryPermissions;
};

export type UserSettings = {
  criticalPathEnabled: boolean;
  displaySummaryTasks: boolean;
};

export type Source = {
  id: number;
  type: string;
};
// ...
export type ProjectSettings = {
  lengthOfDay: number;
  nonWorkingDays: string[];
  workingDays: string[];
};

export type Summary = {
  fields: Field[];
};

export type Field = {
  id: number;
  contactOptions: User[];
  createdAt: string;
  createdBy: User;
  displayValue: string;
  format: string;
  formula: string;
  hyperlink: Hyperlink;
  image: Image;
  index: number;
  locked: boolean;
  lockedForUser: boolean;
  modifiedAt: string;
  modifiedBy: User;
  ObjectValue: ObjectValue;
  options: string[];
  symbol: string;
  title: string;
  type: string;
  validation: boolean;
};
// ...
export type Cell = {
  columnId: number;
  columnType: string;
  conditionalFormat: string;
  displayValue: string;
  format: string;
  formula: string;
  hyperlink: Hyperlink;
  image: Image;
  linkInFromCell: LinkInFromCell;
  linksOutToCells: LinkInFromCell[];
  objectValue: ObjectValue;
  overrideValidation: boolean;
  strict: boolean;
  value: string | number | boolean;
};
// ...
export type Column = {
  autoNumberFormat: AutoNumberFormat;
  contactOptions: User[];
  description: string;
  format: string;
  hidden: boolean;
  id: number;
  index: number;
  locked: boolean;
  lockedForUser: boolean;
  options: string[];
  primary: boolean;
  symbol: string;
  systemColumnType: SystemColumns;
  tags: Tags[];
  title: string;
  type: ColumnTypes;
  validation: boolean;
  version: number;
  width: number;
};
/* ## END COLUMN ## */
/* ############################## */
/* ## COMMENTS ## */
export type Comment = {
  attachments: Attachment[];
  createdAt: string | number;
  createdBy: User;
  discussionId: number;
  id: number;
  modifiedAt: string | number;
  text: string;
};
/* ## END COMMENTS ## */
/* ############################## */
/* ## CROSS SHEET ## */
export type CrossSheetReference = {
  endColumnId: number;
  endRowId: number;
  id: number;
  name: string;
  startColumnId: number;
  startRowId: number;
  status: Status;
  sourceSheetId: number;
};
/* ## END CROSS SHEET ## */
/* ############################## */
/* ## DASHBOARDS ## */
export type Sight = {
  backgroundColor: string;
  columnCount: number;
  favorite: boolean;
  source: Source;
  widgets: Widget[];
  workspace: Workspace;
  createdAt: string | number;
  modifiedAt: string | number;
  id: number;
  accessLevel: AccessLevel;
  permalink: string;
  name: string;
};
// ...
export type Discussion = {
  accessLevel: AccessLevel;
  id: number;
  comments: Comment[];
  commentAttachments: Attachment[];
  commentCount: number;
  createdBy: User;
  lastCommentedAt: string | number;
  lastCommentedUser: User;
  parentId: number;
  parentType: "ROW" | "SHEET";
  readonly: boolean;
  title: string;
};
// ...
export type Folder = {
  id: number;
  favorite: boolean;
  folders: Folder[];
  name: string;
  permalink: string;
  reports: Report[];
  sheets: Sheet[];
  sights: Sight[];
  templates: Template[];
};
// ...
export type Proof = {
  id: number;
  originalId: number;
  name: string;
  proofType: ProofType;
  proofRequestUrl: string;
  version: number;
  lastUpdatedAt: string | number;
  lastUpdatedBy: User;
  isCompleted: boolean;
  attachments: Attachment[];
  discussions: Discussion[];
};

export type ProofRequest = {
  id: number;
  proofId: number;
  sentBy: User;
  sentAt: string | number;
  isDownloadable: boolean;
  status: ProofStatus;
  ccMe: boolean;
  message: string;
  sendTo: User[];
  subject: string;
};

export type ProofRequestAction = {
  user: User;
  actionStatus: ProofStatus;
};
/* ## END PROOFS ## */
/* ############################## */
/* ## REPORTS ## */
export type Report = {
  scope: Scope;
  sourceSheets: Sheet[];
  isSummaryReport: boolean;
  proofs: Proof[];
  id: number;
  fromId: number;
  ownerId: number;
  accessLevel: AccessLevel;
  attachments: Attachment[];
  columns: Column[];
  createdAt: string | number;
  crossSheetReferences: CrossSheetReference[];
  dependenciesEnabled: boolean;
  discussions: Discussion[];
  effectiveAttachmentOptions: string[];
  favorite: boolean;
  ganttEnabled: boolean;
  hasSummaryFields: boolean;
  modifiedAt: string | number;
  name: string;
  owner: string;
  permalink: string;
  projectSettings: ProjectSettings;
  readonly: boolean;
  resourceManagementEnabled: boolean;
  rows: Row[];
  showParentRowsForFilters: boolean;
  source: Source;
  summary: Summary;
  totalRowCount: number;
  userPermissions: UserPermissions;
  userSettings: UserSettings;
  version: number;
  workspace: Workspace;
};
// ...
export type Row = {
  id: number;
  sheetId: number;
  accessLevel: AccessLevel;
  attachments: Attachment[];
  cells: Cell[];
  columns: Column[];
  conditionalFormat: string;
  createdAt: string | number;
  createdBy: User;
  discussions: Discussion[];
  proofs: Proof;
  expanded: boolean;
  filteredOut: boolean;
  format: string;
  inCriticalPath: boolean;
  locked: boolean;
  lockedForUser: boolean;
  modifiedAt: string | number;
  modifiedBy: User;
  permalink: string;
  rowNumber: number;
  version: number;
};
// ...
export type Sheet = {
  id: number;
  fromId: number;
  ownerId: number;
  accessLevel: AccessLevel;
  attachments: Attachment[];
  columns: Column[];
  createdAt: string | number;
  crossSheetReferences: CrossSheetReference[];
  dependenciesEnabled: boolean;
  discussions: Discussion[];
  effectiveAttachmentOptions: string[];
  favorite: boolean;
  ganttEnabled: boolean;
  hasSummaryFields: boolean;
  modifiedAt: string | number;
  name: string;
  owner: string;
  permalink: string;
  projectSettings: ProjectSettings;
  readonly: boolean;
  resourceManagementEnabled: boolean;
  rows: Row[];
  showParentRowsForFilters: boolean;
  source: Source;
  summary: Summary;
  totalRowCount: number;
  userPermissions: UserPermissions;
  userSettings: UserSettings;
  version: number;
  workspace: Workspace;
};

export type Criteria = {
  columnId: number;
  operator: Operators;
  values: string[];
};

export type Filters = {
  id: number;
  excludeSelected: boolean;
  filterType: "PERSONAL" | "SHARED";
  query: Query;
};

export type Query = {
  operator: Operators;
  criteria: Criteria[];
  includeParent: boolean;
};
/* ## END SHEETS ## */
/* ############################## */
/* ## TEMPLATES ## */
export type Template = {
  id: number;
  type: ObjectType;
  accessLevel: AccessLevel;
  blank: boolean;
  categories: string[];
  description: string;
  globalTemplate: "BLANK_SHEET" | "PROJECT_SHEET" | "TASK_LIST";
  image: string;
  largeImage: string;
  locale: string;
  name: string;
  tags: Tags[];
};
/* ## END TEMPLATES ## */
/* ############################## */
/* ## WORKSPACES ## */
export type Workspace = {
  accessLevel: AccessLevel;
  id: number;
  favorite: boolean;
  folders: Folder[];
  name: string;
  permalink: string;
  reports: Report[];
  sheets: Sheet[];
  sights: Sight[];
  templates: Template[];
};